Time: 2497 ms (5.14%) | Memory: 32.3 MB (41.59%) - LeetSync

+++ 1725-number-of-sets-of-k-non-overlapping-line-segments/number-of-sets-of-k-non-overlapping-line-segments.cpp
@@ -7,7 +7,7 @@ public:
         if(k == 0) return 1;
         if(dp[n][k] != -1) return dp[n][k];
 
-        long long result = solve(n-1,k,dp);
+        long result = solve(n-1,k,dp);
         for(int i = 1; i <= m; i++) {
             result = (result + solve(n-i,k-1,dp)) % mod;
         }
